+++ goopylib/src/Platform/Independent/BaseBuffer.h
@@ -97,7 +97,6 @@ namespace gp {
     };
 }
 
-
 // Index Buffer
 namespace gp {
     class BaseIndexBuffer : public BaseBuffer {