+++ goopylib/src/Platform/Independent/BaseShader.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include "pch.h"
+
+
+namespace gp {
+    class BaseShader {
+    public:
+        virtual ~BaseShader() = default;
+
+        virtual void bind() const = 0;
+
+        virtual void unbind() const = 0;
+    };
+}